CPUID: Add a CPUID class to read from /dev/cpu/N/cpuid
Implement the mechanism to read `/dev/cpu/N/cpuid`, where N is the CPU number. On Linux, one should insert the module with `modprobe cpuid`. On QNX, this driver is not implemented yet.

This commit shows in principle how we will handle operating system specifics in the `rjcp::os::qnx::native` namespace, by wrapping direct system calls. As system calls can fail, we gracefully handle them with a proposed C++23 construct with `stdext::expected` taken from [1].

The datatype for 64-bit `lseek64` and `pread64` is documented by [2] to be `off64_t`, even if the headers used shows `loff_t`, so `off64_t` was chosen (and happens to be compatible with QNX also).

Issue: DOTNET-676

[1] https://github.com/TartanLlama/expected
[2] https://linux.die.net/man/3/lseek64