CPUID: Add factory classes for the CpuId* classes
Create factory classes based on templates, which all derive from a single base class `ICpuIdFactory`. This can allow a factory to be instantiated by a template, and given to a non-templated function to create the object required for a specific CPU.

Templates are used for the factory to avoid run-time type checking of the object that should be created. Objects are created based on the templated `CpuId*Config` class. This is why some of the config classes are empty. The specializations are then provided in the `.cpp` files. If the user provides the wrong configuration file, or it isn't implemented, a compiler error will occur (as opposed to run-time, which at best a runtime error occurs, at worst, the programmer forgets and there is undetected, incorrect behaviour).

Note, that the template specialization for the `CpuIdFactory<CpuIdDeviceConfig>` must be in the header file, as it needs an additional private member. If it is not in the header file, but in a CPP file, no warnings or errors will occur at compile time, but runtime may lead to memory corruption because the compiler actually doesn't know what the size of the class is from the header otherwise. This was caught with the Address Sanitizer on Linux.

Issue: DOTNET-676