CpuIdFactory: Simplify the interface
The concrete CpuIdFactory does not need to be exposed in the header for
the users. Instead a single templated free function would suffice to
create a factory parameterized with the desired configuration object.
This unclutters the header and avoids exposing unneeded dependencies in
the header.

By introducing a helper class `CpuIdFactoryWithCallback` for each
concrete factory we only need one specialization of the free function
`CreateCpuIdFactory()`. Each specialization injects a callback in in the
helper class that contains the actual creation logic and captures the
configuration object.

Further refactoring could extract the concrete specializations of
`CreateCpuIdFactory` into separate translation units. This would
completely decouple the concrete types from each other, and save
recompilation of all factories if just one concrete type is changed.