CpuIdFactory: Update design documentation and rationale
One observation with the previous implementation is avoiding a second class of bug that might introduce very hard to find memory bugs. The usage of `CreateCpuIdFactory` free functions, implemented with callbacks and captured lambdas, means that there is always a consistent memory model and impossible to create a class in a header that is different to a class in a CPP file (memory layout) that will result in bugs.

Because it's no longer implemented, the text warning about this class of bug has been removed from the design

Issue: DOTNET-681