CPUID Register: Update structure to keep input values
Later we'll sort the results of the queries in an std::set, which will be sorted on the input values (as opposed to a map, which makes it a bit harder to iterate over).

We don't expect there to be much of a memory increase and thus can deal with having to duplicate the values if needed later.

Issue: DOTNET-678