clang-tidy: Remove magic numbers as a check
Magic numbers are used everywhere in this project, as this is all about low-level system programming of CPU registers and testing. Having the lint generates more exceptions than it's worth and brings little value. Of course we should use constants were possible, but this makes more sense when actually defining the bit fields, instead of just random numbers that we don't care about.