CpuId: Implement simulation factory and test GetCpuId
Create a `CpuIdSimulation` object, generated by a `CpuIdSimulationFactory` (private). The configuration object `CpuIdSimulationConfig` has the `CpuIdTree` that contains all the data which should be returned.

The factory had to be modified to return the number of CPUs, as in a real system, this is the same as `thread::hardware_concurrency()`, but for a simulation, this might be different (e.g. only a single CPU might be available).

Test cases are then implemented using the simulation based on real data recorded on various CPUs. This results in the `GetCpuId()` function being corrected to ensure that the value returned is valid.

Various issues were fixed from the test cases, where CPU data is taken from real-world experiments

Issue: DOTNET-678